+++ Java/src/main/java/localhost/models/UserResponse.java
@@ -39,8 +39,8 @@ public class UserResponse {
     private String type;
     private String parenttype;
     private String tenantId;
-    private Integer indexVersion;
-    private Integer generationNum;
+    private Double indexVersion;
+    private Double generationNum;
     private Double created;
     private Double modified;
     private UserNameAndID author;
@@ -79,8 +79,8 @@ public class UserResponse {
      * @param  type  String value for type.
      * @param  parenttype  String value for parenttype.
      * @param  tenantId  String value for tenantId.
-     * @param  indexVersion  Integer value for indexVersion.
-     * @param  generationNum  Integer value for generationNum.
+     * @param  indexVersion  Double value for indexVersion.
+     * @param  generationNum  Double value for generationNum.
      * @param  created  Double value for created.
      * @param  modified  Double value for modified.
      * @param  author  UserNameAndID value for author.
@@ -112,8 +112,8 @@ public class UserResponse {
             String type,
             String parenttype,
             String tenantId,
-            Integer indexVersion,
-            Integer generationNum,
+            Double indexVersion,
+            Double generationNum,
             Double created,
             Double modified,
             UserNameAndID author,
@@ -656,39 +656,39 @@ public class UserResponse {
 
     /**
      * Getter for IndexVersion.
-     * @return Returns the Integer
+     * @return Returns the Double
      */
     @JsonGetter("indexVersion")
     @JsonInclude(JsonInclude.Include.NON_NULL)
-    public Integer getIndexVersion() {
+    public Double getIndexVersion() {
         return indexVersion;
     }
 
     /**
      * Setter for IndexVersion.
-     * @param indexVersion Value for Integer
+     * @param indexVersion Value for Double
      */
     @JsonSetter("indexVersion")
-    public void setIndexVersion(Integer indexVersion) {
+    public void setIndexVersion(Double indexVersion) {
         this.indexVersion = indexVersion;
     }
 
     /**
      * Getter for GenerationNum.
-     * @return Returns the Integer
+     * @return Returns the Double
      */
     @JsonGetter("generationNum")
     @JsonInclude(JsonInclude.Include.NON_NULL)
-    public Integer getGenerationNum() {
+    public Double getGenerationNum() {
         return generationNum;
     }
 
     /**
      * Setter for GenerationNum.
-     * @param generationNum Value for Integer
+     * @param generationNum Value for Double
      */
     @JsonSetter("generationNum")
-    public void setGenerationNum(Integer generationNum) {
+    public void setGenerationNum(Double generationNum) {
         this.generationNum = generationNum;
     }
 
@@ -881,8 +881,8 @@ public class UserResponse {
         private String type;
         private String parenttype;
         private String tenantId;
-        private Integer indexVersion;
-        private Integer generationNum;
+        private Double indexVersion;
+        private Double generationNum;
         private Double created;
         private Double modified;
         private UserNameAndID author;
@@ -1133,20 +1133,20 @@ public class UserResponse {
 
         /**
          * Setter for indexVersion.
-         * @param  indexVersion  Integer value for indexVersion.
+         * @param  indexVersion  Double value for indexVersion.
          * @return Builder
          */
-        public Builder indexVersion(Integer indexVersion) {
+        public Builder indexVersion(Double indexVersion) {
             this.indexVersion = indexVersion;
             return this;
         }
 
         /**
          * Setter for generationNum.
-         * @param  generationNum  Integer value for generationNum.
+         * @param  generationNum  Double value for generationNum.
          * @return Builder
          */
-        public Builder generationNum(Integer generationNum) {
+        public Builder generationNum(Double generationNum) {
             this.generationNum = generationNum;
             return this;
         }
